refactor: DropHelper -> ConcurrentDroper
