+++ execution/executor/src/components/block_tree/mod.rs
@@ -11,7 +11,7 @@ use crate::logging::{LogEntry, LogSchema};
 use anyhow::{anyhow, ensure, Result};
 use aptos_consensus_types::block::Block as ConsensusBlock;
 use aptos_crypto::HashValue;
-use aptos_drop_helper::DEFAULT_DROP_HELPER;
+use aptos_drop_helper::DEFAULT_DROPPER;
 use aptos_executor_types::{execution_output::ExecutionOutput, ExecutorError, LedgerUpdateOutput};
 use aptos_infallible::Mutex;
 use aptos_logger::{debug, info};
@@ -282,7 +282,7 @@ impl BlockTree {
             old_root
         };
 
-        Ok(DEFAULT_DROP_HELPER.schedule_drop_with_waiter(old_root))
+        Ok(DEFAULT_DROPPER.schedule_drop_with_waiter(old_root))
     }
 
     pub fn add_block(
